refactor: seperate view from service class
